refactor : 일반 근무 관련 조회 코드 정리
