+++ src/main/java/com/example/sidepot/command/employment/app/EmploymentReadService.java
@@ -4,23 +4,15 @@ import com.example.sidepot.command.attendance.domain.Attendance;
 import com.example.sidepot.command.attendance.domain.AttendanceStatus;
 import com.example.sidepot.command.attendance.domain.CoverAttendance;
 import com.example.sidepot.command.attendance.repository.AttendanceDaoRepository;
-import com.example.sidepot.command.attendance.repository.AttendanceRepository;
 import com.example.sidepot.command.attendance.repository.CoverAttendanceDaoRepository;
-import com.example.sidepot.command.attendance.repository.CoverAttendanceRepository;
-import com.example.sidepot.command.employment.domain.Employment;
-import com.example.sidepot.command.employment.domain.EmploymentRepository;
 import com.example.sidepot.command.employment.dto.EmploymentListResDto;
-import com.example.sidepot.command.employment.query.EmploymentDaoRepository;
+import com.example.sidepot.command.employment.repository.EmploymentDaoRepository;
 import com.example.sidepot.command.work.domain.CoverWork;
 import com.example.sidepot.command.work.domain.WorkManager;
 import com.example.sidepot.command.work.domain.WorkTime;
-import com.example.sidepot.command.work.repository.CoverWorkRepository;
 import com.example.sidepot.command.work.repository.WorkManagerRepository;
 import com.example.sidepot.command.work.repository.query.CoverWorkDaoRepository;
-import com.example.sidepot.command.work.repository.query.WorkTimeDaoRepository;
 import lombok.RequiredArgsConstructor;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
