+++ src/main/java/com/example/sidepot/command/employment/presentation/EmploymentCommandController.java
@@ -73,6 +73,4 @@ public class EmploymentCommandController {
                 .data("")
                 .build());
     }
-
-
 }
